Move more backend classes to ble package
