+++ app/src/main/java/edu/drexel/lapcounter/lapcounter/backend/ble/DeviceScanner.java
@@ -1,4 +1,4 @@
-package edu.drexel.lapcounter.lapcounter.backend;
+package edu.drexel.lapcounter.lapcounter.backend.ble;
 
 
 import java.util.List;
