+++ app/src/main/java/edu/drexel/lapcounter/lapcounter/backend/ble/MovingAverage.java
@@ -1,4 +1,4 @@
-package edu.drexel.lapcounter.lapcounter.backend;
+package edu.drexel.lapcounter.lapcounter.backend.ble;
 
 
 /**
