+++ app/src/main/java/edu/drexel/lapcounter/lapcounter/frontend/DeviceSelectActivity.java
@@ -15,7 +15,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import edu.drexel.lapcounter.lapcounter.R;
-import edu.drexel.lapcounter.lapcounter.backend.DeviceScanner;
+import edu.drexel.lapcounter.lapcounter.backend.ble.DeviceScanner;
 import edu.drexel.lapcounter.lapcounter.backend.dummy.DummyDeviceScanner;
 import edu.drexel.lapcounter.lapcounter.frontend.navigationbar.NavBar;
 import edu.drexel.lapcounter.lapcounter.frontend.temp.LapCounterServiceTest;
